Refactor tenant manager tests and add new test cases
- Removed redundant test case for RegisterCustomDomain_PortStripped.
- Added new test case for resolving tenant by custom domain in ResolveContext_ViaCustomDomain.
- Updated test for health check in PostUpdateHealthCheckFunctionalTest to improve clarity.
- Introduced a new log file for Themis tests after filter fix, capturing detailed execution logs.
- Addressed skipped tests in various modules due to platform-specific limitations.
- Fixed assertion error in data augmentation tests to ensure question reformulations are valid.

+++ src/llm/gpu_memory_manager.cpp
@@ -333,8 +333,8 @@ void* GPUMemoryManager::allocateGPU(const std::string& model_id, size_t bytes) {
     std::lock_guard<std::mutex> lock(mutex_);
     
     if (!canAllocate(bytes, 0)) {
-        size_t bytes_mb = bytes / (1024 * 1024);
-        size_t available_mb = (config_.max_vram_bytes - total_vram_used_) / (1024 * 1024);
+        double bytes_mb = static_cast<double>(bytes) / (1024.0 * 1024.0);
+        double available_mb = static_cast<double>(config_.max_vram_bytes - total_vram_used_) / (1024.0 * 1024.0);
         spdlog::error("[{}] GPU OOM: requested {:.1f} MB, available {:.1f} MB", 
                       static_cast<int>(errors::ErrorCode::ERR_LLM_GPU_OOM), 
                       bytes_mb, available_mb);
